refactor-#29: adjust infrastructure module
