+++ src/customer/application/application.module.ts
@@ -2,7 +2,7 @@ import { Global, Module } from '@nestjs/common';
 import { CqrsModule } from '@nestjs/cqrs';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { CustomerRepository } from '../repositories';
+import { CustomerRepository } from '../infrastructure/repositories';
 import { CommandHandlers } from './commands/handlers';
 import { CustomerController } from './controllers';
 import { QueryHandlers } from './queries/handlers';
