+++ src/customer/infrastructure/infrastructure.module.ts
@@ -1,4 +1,11 @@
-import { Module } from '@nestjs/common';
+import { Global, Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
 
-@Module({})
+import { CustomerEntity } from './entities';
+import { CustomerRepository } from './repositories';
+
+@Global()
+@Module({
+    imports: [TypeOrmModule.forFeature([CustomerRepository, CustomerEntity])]
+})
 export class InfrastructureModule {}
